Event queue: add event queue implementation for monitoring fds
Adds a kqueue/epoll like interface for monitoring file descriptors and
other objects.

event_queue_create() creates a queue and returns an fd for the queue
event_queue_select() adds, removes and modifies monitored objects
event_queue_wait() waits for events on the queue

Events are currently edge-triggered only. When an object is deleted,
a B_EVENT_INVALID event is delivered, and the object is unregistered
from the queue.
The special event flag B_EVENT_ONE_SHOT can be passed in when adding
an object so that the object is automatically unregistered when an
event is delivered.